Move result extraction logic to callback object
