--- graphql-native/src/main/java/io/ballerina/stdlib/graphql/runtime/utils/CallableUnitCallback.java
@@ -1,52 +1,0 @@
-/*
- * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package io.ballerina.stdlib.graphql.runtime.utils;
-
-import io.ballerina.runtime.api.async.Callback;
-import io.ballerina.runtime.api.values.BError;
-
-import java.util.concurrent.CountDownLatch;
-
-/**
- * This class is used as a callback class for Ballerina resource execution.
- */
-public class CallableUnitCallback implements Callback {
-    private CountDownLatch latch;
-    private Object result;
-
-    public CallableUnitCallback(CountDownLatch latch) {
-        this.latch = latch;
-    }
-
-    public Object getResult() {
-        return this.result;
-    }
-
-    @Override
-    public void notifySuccess(Object o) {
-        this.result = o;
-        this.latch.countDown();
-    }
-
-    @Override
-    public void notifyFailure(BError bError) {
-        this.result = bError;
-        this.latch.countDown();
-    }
-}
